+++ water/.prettierignore
@@ -1,0 +1,5 @@
+/*
+!/src
+!eslintrc.cjs
+!package.json
+!tsconfig.json
