+++ water/.prettierrc.yml
@@ -1,0 +1,10 @@
+useTabs: true
+singleQuote: false
+jsxSingleQuote: false
+bracketSameLine: true
+quoteProps: as-needed
+trailingComma: es5
+semi: true
+printWidth: 120
+arrowParens: avoid
+endOfLine: auto
