+++ water/package.json
@@ -1,27 +1,29 @@
 {
-    "name": "@beemobot/water",
-    "version": "1.0.0",
-    "author": "Beemo Devs (https://beemo.gg)",
-    "license": "MIT",
-    "type": "module",
-    "main": "./dist/index.js",
-    "types": "./dist/index.d.ts",
-    "scripts": {
-        "build": "rm -rf dist/ && tsc",
-        "lint": "eslint ."
-    },
-    "files": [
-        "dist"
-    ],
-    "dependencies": {
-        "kafkajs": "^2.2.4"
-    },
-    "devDependencies": {
-        "@types/node": "^20.11.19",
-        "@typescript-eslint/eslint-plugin": "^7.0.1",
-        "@typescript-eslint/parser": "^7.0.1",
-        "eslint": "^8.56.0",
-        "eslint-plugin-import": "^2.29.1",
-        "typescript": "^5.3.3"
-    }
+	"name": "@beemobot/water",
+	"version": "1.0.0",
+	"author": "Beemo Devs (https://beemo.gg)",
+	"license": "MIT",
+	"type": "module",
+	"main": "./dist/index.js",
+	"types": "./dist/index.d.ts",
+	"scripts": {
+		"build": "rm -rf dist/ && tsc",
+		"format": "prettier --write .",
+		"lint": "eslint ."
+	},
+	"files": [
+		"dist"
+	],
+	"dependencies": {
+		"kafkajs": "^2.2.4"
+	},
+	"devDependencies": {
+		"@types/node": "^20.11.19",
+		"@typescript-eslint/eslint-plugin": "^7.0.1",
+		"@typescript-eslint/parser": "^7.0.1",
+		"eslint": "^8.56.0",
+		"eslint-plugin-import": "^2.29.1",
+		"prettier": "^3.2.5",
+		"typescript": "^5.3.3"
+	}
 }
