+++ water/src/broker/IBrokerMessageHeaders.ts
@@ -1,15 +1,13 @@
 export const BROKER_HEADERS = {
-    CLIENT_ID: "client-id",
-    REQUEST_ID: "request-id",
-    SOURCE_CLUSTER: "source-cluster",
-    TARGET_CLUSTERS: "target-clusters",
+	CLIENT_ID: "client-id",
+	REQUEST_ID: "request-id",
+	SOURCE_CLUSTER: "source-cluster",
+	TARGET_CLUSTERS: "target-clusters",
 };
 
 export interface IBrokerMessageHeaders {
-
-    get clientId(): string;
-    get sourceCluster(): string;
-    get targetClusters(): Set<string>;
-    get requestId(): string;
-
+	get clientId(): string;
+	get sourceCluster(): string;
+	get targetClusters(): Set<string>;
+	get requestId(): string;
 }
