+++ water/src/index.ts
@@ -2,8 +2,8 @@ export * from "./broker/BrokerClient.js";
 export * from "./broker/BrokerMessage.js";
 export * from "./broker/IBrokerConnection.js";
 export * from "./broker/IBrokerMessageHeaders.js";
-export * from "./broker/KafkaConnection.js";
-export * from "./broker/KafkaMessageHeaders.js";
+export * from "./broker/kafka/KafkaConnection.js";
+export * from "./broker/kafka/KafkaMessageHeaders.js";
 
 export * from "./logging/Logger.js";
