+++ water/src/util/CountDownLatch.ts
@@ -1,7 +1,6 @@
 type PromiseResolver = (timeout: boolean) => void;
 
 export class CountDownLatch {
-
     private counter: number;
     private resolvers = new Map<PromiseResolver, NodeJS.Timeout | null>();
 
@@ -42,20 +41,17 @@ export class CountDownLatch {
         }
 
         let resolver: PromiseResolver;
-        const promise = new Promise<boolean>(r => {
-            resolver = r;
-            this.resolvers.set(r, null);
-        });
-        if (!waitTime) {
-            return promise;
-        }
 
-        const timeout = setTimeout(() => {
+        const timeout = waitTime ? setTimeout(() => {
             this.resolvers.delete(resolver);
             resolver(false);
-        }, waitTime);
-        timeout.unref();
+        }, waitTime) : null;
+        timeout?.unref();
+
+        const promise = new Promise<boolean>(r => {
+            resolver = r;
+            this.resolvers.set(r, timeout);
+        });
         return promise;
     }
-
 }
