+++ water/tsconfig.json
@@ -1,30 +1,26 @@
 {
-    "compilerOptions": {
-        "target": "ES2022",
-        "module": "Node16",
-        "lib": [
-            "ES2022"
-        ],
-        "moduleResolution": "Node16",
-        "allowJs": false,
-        "allowSyntheticDefaultImports": true,
-        "alwaysStrict": true,
-        "declaration": true,
-        "esModuleInterop": false,
-        "forceConsistentCasingInFileNames": true,
-        "noImplicitAny": true,
-        "noImplicitOverride": true,
-        "noImplicitReturns": true,
-        "noUncheckedIndexedAccess": true,
-        "resolveJsonModule": true,
-        "preserveConstEnums": true,
-        "skipLibCheck": true,
-        "sourceMap": true,
-        "strict": true,
-        "useDefineForClassFields": true,
-        "outDir": "dist"
-    },
-    "include": [
-        "src/**/*"
-    ]
+	"compilerOptions": {
+		"target": "ES2022",
+		"module": "Node16",
+		"lib": ["ES2022"],
+		"moduleResolution": "Node16",
+		"allowJs": false,
+		"allowSyntheticDefaultImports": true,
+		"alwaysStrict": true,
+		"declaration": true,
+		"esModuleInterop": false,
+		"forceConsistentCasingInFileNames": true,
+		"noImplicitAny": true,
+		"noImplicitOverride": true,
+		"noImplicitReturns": true,
+		"noUncheckedIndexedAccess": true,
+		"resolveJsonModule": true,
+		"preserveConstEnums": true,
+		"skipLibCheck": true,
+		"sourceMap": true,
+		"strict": true,
+		"useDefineForClassFields": true,
+		"outDir": "dist"
+	},
+	"include": ["src/**/*"]
 }
